pyproject.toml
[tool.poetry]
name = "osv-dev-api-client-go"
version = "0.1.0"
description = ""
authors = ["Paul Horton <paul.horton@owasp.org>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
